Vendor github.com/hashicorp/golang-lru
This is part of an experiment in using a LRU for an open
object handle cache. The upstream implementation of TwoQueueCache
doesn't provide a way to run a callback on remove/evict, so we'll
need to hack it a bit. Hopefully upstream will accept our
modifications so we don't need to maintain this ourselves.