Support for specifying an epoch for nodes
If a daosfs.Node readEpoch is set, then that node
will read all data from DAOS using that Epoch, as if
it were part of a snapshot.

This value can be set using the "user.snapshot_epoch"
extended attribute. If this attribute is set on a file,
then it will be in "snapshot" mode and udpates to
the file will faile with EPERM errors. Setting this
value thorugh xattr is temporary for testing and
should be removed once snapshots are implemented.