Rework filehandles
The Write() implementation was getting a new tx for every call
to Write(), which doesn't make sense. What we want is to get a tx
when the FH is opened for write and then commit the tx when the file
is being closed.

NB: Still need to handle fsync()/flush() cases.