Major refactoring of pkg/daos
Aims to clean up the core package by removing a lot of Go alias types
for DAOS stuff. The goal is to reduce cognitive burden for thinking
about what's happening in DAOS vs. what's happening in Go. Also
simplify a lot of operations to use a new IoRequest type which
deals with getting keys and buffers into DAOS-compatible shapes.